% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saw_internal.R
\name{compute_threshold}
\alias{compute_threshold}
\title{Compute threshold parameter.}
\usage{
compute_threshold(y, x, N, TT, PP, s_thresh, gamma_tilde, kappa = NULL)
}
\description{
This function computes the threshold parameter used to shrink the \code{b}
coefficients to zero. The parameter is found in the paper under the label
\verb{\\lambda_\{NT\}}.
}
