% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{compute_w_matrix_and_b_tilde}
\alias{compute_w_matrix_and_b_tilde}
\title{Computes the \verb{W_\{l,k\}} and \verb{\\tilde\{b\}_\{l,k\}} given index \verb{\{l,k\}}.}
\usage{
compute_w_matrix_and_b_tilde(index, delta_y, x, z, wavelet_bases, TT, N, PP)
}
\description{
Computes the \verb{W_\{l,k\}} and \verb{\\tilde\{b\}_\{l,k\}} given index \verb{\{l,k\}}.
}
