% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{construct_a_matrices}
\alias{construct_a_matrices}
\title{Construct \code{A} matrices.}
\usage{
construct_a_matrices(index, Q, PP)
}
\description{
These matrices can be found in section 2.3 of the paper.
}
\details{
A_minus corresponds to \verb{A_\{l, 2k - 1\}} while A corresponds to \verb{A_\{l, 2k\}}
with the exception of index == 1 where A corresponds to \verb{A_\{1, 1\}}.

For parameter \code{Q} check function \code{construct_q_matrices}.
}
