% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{construct_h_matrices}
\alias{construct_h_matrices}
\title{Construct \verb{H_\{l, 2k\}, H_\{l, 2k -1 \}} matrices.}
\usage{
construct_h_matrices(index, wavelet_bases, TT)
}
\description{
The definition of these matrices is found in section 2.2 of the paper.
}
