% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{construct_q_matrices}
\alias{construct_q_matrices}
\title{Construct \verb{\\underscore\{Q\}} matrices.}
\usage{
construct_q_matrices(index, x, z, N, TT, selector_minus, selector, H)
}
\description{
These matrices can be found in section 2.3 of the paper.
}
\details{
Q_minus corresponds to \verb{\\underscore\{Q\}_\{l, 2k-1\}} while Q, except for the
case index == 1, corresponds to \verb{\\underscore\{Q\}_\{l, 2k\}}.

For parameter \code{H} check function \code{construct_h_matrices}.
}
