% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saw_internal.R
\name{data_to_matrix}
\alias{data_to_matrix}
\title{Construct data matrix where first column contains \code{delta_y} and other
columns correspond to \verb{\\underscore\{X\}}. The reference can be found in the
paper in section 2.1.}
\usage{
data_to_matrix(delta_y, shifted_x, unshifted_x, P)
}
\description{
Construct data matrix where first column contains \code{delta_y} and other
columns correspond to \verb{\\underscore\{X\}}. The reference can be found in the
paper in section 2.1.
}
