% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saw_internal.R
\name{feature_list_to_matrix}
\alias{feature_list_to_matrix}
\title{Convert list of P (TxN) matrices to single matrix with shape T x (N * P).
Data is stored as \link{"x_{t, 11}, x_{t, 12}, ..., x_{t, PN}"}.}
\usage{
feature_list_to_matrix(X)
}
\description{
Convert list of P (TxN) matrices to single matrix with shape T x (N * P).
Data is stored as \link{"x_{t, 11}, x_{t, 12}, ..., x_{t, PN}"}.
}
