% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_saw.R
\name{fit_saw_iter}
\alias{fit_saw_iter}
\title{Iterative threshold finder}
\usage{
fit_saw_iter(
  y,
  X,
  Z = NULL,
  time_effect = TRUE,
  id_effect = TRUE,
  max_iter = 20,
  return_info = FALSE
)
}
\arguments{
\item{y}{Matrix of labels. Has dimension T x N.}

\item{X}{List of feature matrices. The pth entry corresponds to the design
matrix of the pth covariate and has dimension T x N.}

\item{Z}{Instruments corresponding to the argument X. If NULL all X variables
are their own instrument.}

\item{time_effect}{Boolean indicating if a time effect is to be estimated.}

\item{id_effect}{Boolean indicating if an individual effect is to be
estimated and returned.}

\item{max_iter}{Maximum number of iterations. Default 20.}

\item{return_info}{Return additional info on model fit.}
}
\description{
Solve for threshold using iterative procedure which sets the
threshold as the smallest absolute coefficient in iteration until convergence.
}
