% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_saw.R
\name{predict.saw_model}
\alias{predict.saw_model}
\title{Predict with a fitted saw model}
\usage{
\method{predict}{saw_model}(model, X)
}
\arguments{
\item{model}{fitted saw model output from function fit_saw}

\item{X}{newdata of same shape as the one used to fit model}
}
\description{
Predict with a fitted saw model
}
