% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saw_internal.R
\name{repmat}
\alias{repmat}
\title{Repeat matrix in the both axis \code{rows} and \code{cols} times.}
\usage{
repmat(mat, rows = 1, cols = 1)
}
\description{
Repeat matrix in the both axis \code{rows} and \code{cols} times.
}
