% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{robust_matrix_square_root}
\alias{robust_matrix_square_root}
\title{Compute the square root of \code{mat}.}
\usage{
robust_matrix_square_root(mat)
}
\description{
This method also works in the case where \code{mat} is non-symmetric but is
not accurate in this case. For more information seek: https://tinyurl.com/97ek4f7u
}
