% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_MAIN.R
\name{saw_fun}
\alias{saw_fun}
\title{SAW Estimation Procedure}
\usage{
saw_fun(formula, dot = FALSE, s.thresh = NULL, ridge = 1e-04)
}
\arguments{
\item{formula}{A formula object; Note that "-1" is unnecessary since the method takes first differences.}

\item{dot}{A boolean indicating wether a common time trend is to be eliminated.}

\item{s.thresh}{A tuning parameter.}

\item{ridge}{A stabilization parameter to avoid inverting (near) singular design matrices (t(X)X)}
}
\description{
SAW Estimation Procedure
}
