% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saw_internal.R
\name{transform_input_data}
\alias{transform_input_data}
\title{Transform input data which is either a T x N matrix (y) or a list of T x N
matrices (X and Z) to data objects expected from the saw method. That is
we construct \verb{\\delta y} from y and \verb{\\underscore\{X\}}, \verb{\\underscore\{Z\}} from
X and Z, respectively. See section 2.1 in the paper.}
\usage{
transform_input_data(y, X, Z)
}
\description{
Transform input data which is either a T x N matrix (y) or a list of T x N
matrices (X and Z) to data objects expected from the saw method. That is
we construct \verb{\\delta y} from y and \verb{\\underscore\{X\}}, \verb{\\underscore\{Z\}} from
X and Z, respectively. See section 2.1 in the paper.
}
