% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_saw.R
\name{fit_saw_cv}
\alias{fit_saw_cv}
\title{k-fold cross validation to find optimal threshold parameter}
\usage{
fit_saw_cv(
  y,
  X,
  Z = NULL,
  time_effect = FALSE,
  n_folds = 4,
  n_loops = 20,
  parallel = TRUE,
  n_cores = NULL,
  return_info = FALSE
)
}
\arguments{
\item{y}{Matrix of labels. Has dimension T x N.}

\item{X}{List of feature matrices. The pth entry corresponds to the design
matrix of the pth covariate and has dimension T x N.}

\item{Z}{Instruments corresponding to the argument X. If NULL all X variables
are their own instrument.}

\item{time_effect}{Boolean indicating if a time effect is to be estimated.}

\item{n_folds}{Number of folds to use in cross validation step. Default 4. It
makes sense that n_folds = k * n_cores.}

\item{n_loops}{Number of s_thresh candidates.}

\item{parallel}{Boolean indicating if code is parallelized over folds.}

\item{n_cores}{How many cores to use for paralellization. Default to number
of available logical cores. It makes sense that n_folds = k * n_cores.}

\item{return_info}{Return additional info on model fit.}
}
\description{
k-fold cross validation to find optimal threshold parameter
}
