% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_saw.R
\name{saw_fun}
\alias{saw_fun}
\title{Post-SAW Estimation Procedure.}
\usage{
saw_fun(
  y,
  X,
  Z = NULL,
  time_effect = FALSE,
  id_effect = FALSE,
  s.thresh = NULL
)
}
\arguments{
\item{y}{Matrix of labels. Has dimension T x N.}

\item{X}{List of feature matrices. The pth entry corresponds to the design
matrix of the pth covariate and has dimension T x N.}

\item{Z}{Instruments corresponding to the argument X. If NULL all X variables
are their own instrument.}

\item{time_effect}{Boolean indicating if the method needs to control for a
common time trend. This trend will then be estimated and returned.}

\item{id_effect}{Boolean indicating if an individual effect is to be
estimated and returned.}

\item{s.thresh}{Tuning parameter for the threshold lambda.}
}
\description{
Exports main function of the package, which performs the SAW procedure and
afterwards the Post-SAW estimation.
}
