% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_MAIN.R
\name{saw_fun}
\alias{saw_fun}
\title{SAW Estimation Procedure}
\usage{
saw_fun(formula, dot = FALSE, s.thresh = NULL)
}
\arguments{
\item{formula}{A formula object; Note that "-1" is unnecessary since the method takes first differences.}

\item{s.thresh}{A tuning parameter.}

\item{data}{A dataframe containing the variables.}
}
\description{
SAW Estimation Procedure
}
