% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{saw_fun}
\alias{saw_fun}
\title{SAW Estimation Procedure}
\usage{
saw_fun(formula, Z = NULL, time_effect = FALSE, s.thresh = NULL)
}
\arguments{
\item{formula}{Formula object. Intercept is eliminated by taking first
differences.}

\item{Z}{Instruments corresponding to X matrix in formula object. If NULL all
X variables are their own instrument.}

\item{time_effect}{Boolean indicating if the method needs to control for a
common time trend. This trend will then be estimated and returned.}

\item{s.thresh}{Tuning parameter for the threshold lambda.}
}
\description{
SAW Estimation Procedure
}
