% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_MAIN.R
\name{saw_fun}
\alias{saw_fun}
\title{SAW Estimation Procedure}
\usage{
saw_fun(formula, data, s.thresh = NULL)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A dataframe containing the variables.}

\item{s.thresh}{A tuning parameter.}
}
\description{
SAW Estimation Procedure
}
