% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_saw.R
\docType{package}
\name{sawr}
\alias{sawr}
\title{sawr}
\description{
This package contains the implementations of the panel data estimation
procedure suggested in Bada et al. (2020), titled 'A Wavelet Method for Panel
Models with Jump Discontinuities in the Parameters'
}
\references{
Bada et al. (2020). A Wavelet Method for Panel Models with Jump
Discontinuities in the Parameters'
}
